Finished Basic Button Logic
The button now acts as a detection zone. It detects if any object stays inside it's trigger, or if any object leaves it's trigger. This makes it so an object has to be left inside of it in order to stay on, although this should be specified to be objects like the player or the crate. Will work on implementing this with James' logic next